# ui-new/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/data/capability": {
      "get": {
        "tags": [],
        "operationId": "get_capability_history_doc",
        "parameters": [
          {
            "name": "entityId",
            "in": "query",
            "description": "Entity id with capability suffix",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Start timestamp ISO-8601",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "End timestamp ISO-8601",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pagesize",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "description": "Event type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Capability history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CapabilityHistoryDoc"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapabilityHistoryDoc": {
        "type": "object",
        "required": [
          "eventType",
          "eventTime",
          "dataName",
          "dataValue",
          "entityId"
        ],
        "properties": {
          "dataName": {
            "type": "string"
          },
          "dataValue": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "eventTime": {
            "type": "string"
          },
          "eventType": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}